Understanding of setup
App.js has a function called App that returns a jsx which is rendered in the index.js file. In index.js file '<App/>' is a jsx file which is a mixture of html tag and js. The index.js file will be the starting part of the app, the code in the index.js file will be the first code that is executed in the browser when we load our project in the browser. 'import ReactDOM from 'react-dom';'-ReactDOM object is being imported from the 'react-dom' library. The ReactDOM is part of our project because it's part of the dependencies listed in package.json file