Add card, button and style. Understanding function of index.js
In index.js file there is a ‘render’ object that renders the jsx component called <App/>.  Jsx is a special syntax that is not understood by the browser that is converted behind the scenes but it helps for the developers to build components to describe the desired output. In the public folder there is an index.html file which is fetched from the server with a <div id=root></div> that is able to render the components to the browser, if code is looked at in the browser the output will be displayed between the beginning and closing div only showing the content of the component when it runs in the browser.