using props
In order to add data into a component or make use of each component in an App.js file you use props. props stand for property and props is put in the parameters of the Todo function then you put brackets and then put props with a . and text. You can also add dynamic expression. You can not have any block statements in the brackets, block statements like if statements