onClick handler in Todo file
When you put an onClick as an attribute you are telling react to react when the button is clicked. In react the <button> element is a component. The value you assign to the onClick attribute is a dynamic expression that should hold function that should be execute when a click happens. There are 3 ways 1. an inline function, 2. you point and define the function under the todo function with its props  3. you can use an arrow function '() =>'. With a pointer function you don't execute the function ex: {deletHandler()} you do it like this {deleteHandler} you want to execute the f unction when it is clicked, it will execute when java evalutes the code if () is added